+++ lib/nicetest.rb
@@ -8,9 +8,9 @@ require_relative "nicetest/logger"
 require_relative "nicetest/cli"
 require_relative "nicetest/backtrace_filter"
 require_relative "nicetest/test_finder"
-require_relative "nicetest/nicetest_plugin"
-require_relative "nicetest/reporters_plugin"
-require_relative "nicetest/superdiff_plugin"
+require_relative "minitest/nicetest_plugin"
+require_relative "minitest/reporters_plugin"
+require_relative "minitest/superdiff_plugin"
 
 module Nicetest
   class Error < StandardError; end
