+++ lib/nicetest/cli.rb
@@ -11,7 +11,6 @@ module Nicetest
     def initialize(argv)
       @argv = argv
       @logger = Logger.new($stderr)
-      register_minitest_plugins!
       adjust_load_path!
     end
 
@@ -199,24 +198,6 @@ module Nicetest
         define_method(:callback!, original_callback)
         remove_method(:noop_callback!)
       end
-    end
-
-    def register_minitest_plugins!
-      if Minitest.respond_to?(:register_plugin)
-        Minitest.register_plugin(Nicetest::NicetestPlugin)
-        Minitest.register_plugin(Nicetest::ReportersPlugin)
-        Minitest.register_plugin(Nicetest::SuperdiffPlugin)
-      else
-        shim_plugin!(Nicetest::NicetestPlugin, :nicetest)
-        shim_plugin!(Nicetest::ReportersPlugin, :reporters)
-        shim_plugin!(Nicetest::SuperdiffPlugin, :superdiff)
-      end
-    end
-
-    def shim_plugin!(plugin, name)
-      Minitest.define_singleton_method("plugin_#{name}_init", plugin.method(:minitest_plugin_init).to_proc)
-      Minitest.define_singleton_method("plugin_#{name}_options", plugin.method(:minitest_plugin_options).to_proc)
-      Minitest.instance_variable_get(:@extensions) << name
     end
 
     Opts = Struct.new(:cd, :name) do
